Demo project

+++ Banking/src/Model/Transfer.java
@@ -22,21 +22,21 @@ public class Transfer {
         this.daoCustomer = daoCustomer;
     }
     
-    public void transfer(CustomerAccount my, CustomerAccount target, double amount){
+    public void transfer(CustomerAccount my, CustomerAccount target, int amount){
         updateCustomer(my, amount*(-1));
         updateCustomer(target, amount);
         addTransaction(my, amount*(-1));
         addTransaction(target, amount);
     }
     
-    private void updateCustomer(CustomerAccount customer, double amount){
+    private void updateCustomer(CustomerAccount customer, int amount){
         int id = Integer.parseInt(customer.getID().substring(1));
         double balance = customer.getBalance() + amount;
         CustomerAccount updateCustomer = new CustomerAccount(customer, balance);
         daoCustomer.updateCustomerBalance(updateCustomer);
     }
     
-    private void addTransaction(CustomerAccount customer, double amount){
+    private void addTransaction(CustomerAccount customer, int amount){
         int id = Integer.parseInt(customer.getID().substring(1));
         double balance = customer.getBalance() + amount;
         String date = new Timestamp((new Date()).getTime()).toString();
